Move authenticator into it's own module
